修復距離錯誤
將API解析的Int(距離)修改成Double(距離)

+++ WhereIsUBike/Model/BikeStationData.swift
@@ -13,14 +13,14 @@ struct BikeStationData {
     var station_Borrow = ""
     var station_Return = ""
     var station_Distance = ""
-    var station_Distance_Number = 0
+    var station_Distance_Number = 0.0
     var station_Latitude = ""
     var station_Longtitude = ""
     var station_Address = ""
     var userLocation_Latitude: Double = 0.0
     var userLocation_Longtitude: Double = 0.0
     
-    init(station_Title: String,station_Borrow: String,station_Return: String,station_Distance: String,station_Distance_Number: Int,station_Latitude: String,station_Address: String,station_Longtitude: String,userLocation_Latitude: Double,userLocation_Longtitude: Double){
+    init(station_Title: String,station_Borrow: String,station_Return: String,station_Distance: String,station_Distance_Number: Double,station_Latitude: String,station_Address: String,station_Longtitude: String,userLocation_Latitude: Double,userLocation_Longtitude: Double){
         self.station_Title = station_Title
         self.station_Borrow = station_Borrow
         self.station_Return = station_Borrow
